requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bandit==1.7.1
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
Flask==2.0.2
Flask-SQLAlchemy==2.1
gitdb==4.0.9
GitPython==3.1.26
greenlet==1.1.2
idna==3.3
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pylint==2.12.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
responses==0.18.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.31
stack-data==0.1.4
stevedore==3.5.0
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
